Split WithState in to two interfaces so it can apply everywhere, reduce the command one to get the same result
Before it was string?, which can't apply to string ones because you can't set them to null.
Instead I...StateSetter says it's not null (you can't set it to null), and I...StateGetter says it could be null.
These can apply to string? and string ones.
Command probably wants the same thing done to it